requirements.txt
fastapi
uvicorn
sqlmodel
python-multipart
jinja2
python-dotenv
openai
stripe
markitdown
pydantic
python-jose
passlib
bcrypt
psycopg2-binary
pypdf
aiofiles
python-magic
pytest
httpx